auth_app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.2.0"
description = "Authorization Service for a Cinema Online"
authors = [
    "Alexander Oorzhak",
    "Vitaly Shchurov"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = {extras = ["async"], version = "^2.2.2"}
flake8 = "^6.0.0"
flask-sqlalchemy = "^3.0.2"
psycopg2-binary = "^2.9.5"
gevent = "^22.10.2"
configorm = "^0.1.1"
gunicorn = "^20.1.0"
flask-migrate = "^4.0.0"
flask-swagger-ui = "^4.11.1"
marshmallow-sqlalchemy = "^0.28.1"
pyjwt = "^2.6.0"
redis = "^4.4.0"
flask-jwt-extended = "^4.4.4"
flask-security = "^3.0.0"
pydantic = {extras = ["email"], version = "^1.10.2"}
backoff = "^2.2.1"
python-dotenv = "^0.21.0"
device-detector = "^5.0.1"
flask-pydantic = "^0.11.0"
orjson = "^3.8.3"
aioredis = "1.3.1"
factory-boy = "3.2.1"
faker = "15.3.2"
pytest = "6.2.5"
pytest-asyncio = "0.20.2"
pytest-factoryboy = "2.5.0"
aiohttp = "3.8.3"
opentelemetry-api = "1.10.0"
opentelemetry-sdk = "1.10.0"
opentelemetry-instrumentation-flask = "0.29b1"
opentelemetry-exporter-jaeger = "1.10.0"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.11.4"
elasticsearch = {version = "7.9.1", extras = ["async"]}
aioredis = "1.3.1"
fastapi = "0.61.1"
uvicorn = "0.19.0"
uvloop = "0.17.0"
httptools = "0.5.0"
authlib = "^1.2.0"
google-api-python-client = "^2.71.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
